+++ ctor/tests/errors/ctor_no_unsafe_struct.rs
@@ -1,0 +1,13 @@
+use ctor::ctor;
+
+struct FooGeneric<T> {
+    _t: ::std::marker::PhantomData<T>,
+}
+
+impl<T: Default> FooGeneric<T> {
+    #[ctor]
+    fn foo() {
+    }
+}
+
+fn main() {}