+++ ctor/tests/expand-linux/ctor_attribute.rs
@@ -1,6 +1,6 @@
 use ctor::ctor;
 
-#[ctor(link_section = ".ctors")]
+#[ctor(unsafe,link_section = ".ctors")]
 fn foo() {
     println!("foo");
 }