+++ ctor/tests/expand-linux/ctor_warn.rs
@@ -1,6 +1,0 @@
-use ctor::ctor;
-
-#[ctor]
-fn foo() {
-    println!("foo");
-}