+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - `#[ctor(priority = naked)]` is now `#[ctor(naked)]`.
+- `unsafe` is now required for `#[ctor]` items and the
+  `no_warn_on_missing_unsafe` feature is gone.
+  - `RUSTFLAGS="--cfg no_fail_on_missing_unsafe"` can bypass the error.
 
 ### Added